+ Fixed 0000125: Less than 60 minutes are not plotting in the graph
git-svn-id: http://agilefant.googlecode.com/svn/trunk@3963 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/util/MinorUnitsParser.java
@@ -104,7 +104,7 @@ public class MinorUnitsParser {
     }
     
     public double toMajorUnits(long minorUnits) {
-        return minorUnits / minorsPerMajor;
+        return (double)minorUnits / (double)minorsPerMajor;
     }
 
 }
